Fixes #4

+++ src/RWFileCache.php
@@ -134,6 +134,10 @@ class RWFileCache
     {
         $filePath = $this->getFilePathFromKey($key);
 
+        if (!file_exists($filePath)) {
+            return false;
+        }
+
         return unlink($filePath);
     }
